Polish cargo test arguments + add test
