+++ kani-driver/src/args/mod.rs
@@ -3,6 +3,7 @@
 //! Module that define Kani's command line interface. This includes all subcommands.
 
 pub mod assess_args;
+pub mod cargo;
 pub mod common;
 pub mod playback_args;
 
@@ -10,6 +11,7 @@ pub use assess_args::*;
 
 use self::common::*;
 use crate::util::warning;
+use cargo::CargoCommonArgs;
 use clap::builder::{PossibleValue, TypedValueParser};
 use clap::{error::ContextKind, error::ContextValue, error::Error, error::ErrorKind, ValueEnum};
 use kani_metadata::CbmcSolver;
@@ -295,7 +297,7 @@ pub struct VerificationArgs {
 
     /// Arguments to pass down to Cargo
     #[command(flatten)]
-    pub cargo: CargoArgs,
+    pub cargo: CargoCommonArgs,
 
     #[command(flatten)]
     pub common_args: CommonArgs,
@@ -328,93 +330,6 @@ impl VerificationArgs {
             Some(None) => None,       // -j
             Some(Some(x)) => Some(x), // -j=x
         }
-    }
-}
-
-/// Arguments that Kani pass down into Cargo essentially uninterpreted.
-/// These generally have to do with selection of packages or activation of features.
-/// These do not (currently) include cargo args that kani pays special attention to:
-/// for instance, we keep `--tests` and `--target-dir` elsewhere.
-#[derive(Debug, Default, clap::Args)]
-pub struct CargoArgs {
-    /// Activate all package features
-    #[arg(long)]
-    pub all_features: bool,
-    /// Do not activate the `default` feature
-    #[arg(long)]
-    pub no_default_features: bool,
-
-    // This tolerates spaces too, but we say "comma" only because this is the least error-prone approach...
-    /// Comma separated list of package features to activate
-    #[arg(short = 'F', long)]
-    features: Vec<String>,
-
-    /// Path to Cargo.toml
-    #[arg(long, name = "PATH")]
-    pub manifest_path: Option<PathBuf>,
-
-    /// Build all packages in the workspace
-    #[arg(long)]
-    pub workspace: bool,
-
-    /// Run Kani on the specified packages.
-    #[arg(long, short, conflicts_with("workspace"), num_args(1..))]
-    pub package: Vec<String>,
-
-    /// Exclude the specified packages
-    #[arg(long, short, requires("workspace"), conflicts_with("package"), num_args(1..))]
-    pub exclude: Vec<String>,
-}
-
-impl CargoArgs {
-    /// Parse the string we're given into a list of feature names
-    ///
-    /// clap can't do this for us because it accepts multiple different delimeters
-    pub fn features(&self) -> Vec<String> {
-        let mut result = Vec::new();
-
-        for s in &self.features {
-            for piece in s.split(&[' ', ',']) {
-                result.push(piece.to_owned());
-            }
-        }
-        result
-    }
-
-    /// Convert the arguments back to a format that cargo can understand.
-    /// Note that the `exclude` option requires special processing and it's not included here.
-    pub fn to_cargo_args(&self) -> Vec<OsString> {
-        let mut cargo_args: Vec<OsString> = vec![];
-        if self.all_features {
-            cargo_args.push("--all-features".into());
-        }
-
-        if self.no_default_features {
-            cargo_args.push("--no-default-features".into());
-        }
-
-        let features = self.features();
-        if !features.is_empty() {
-            cargo_args.push(format!("--features={}", features.join(",")).into());
-        }
-
-        if let Some(path) = &self.manifest_path {
-            cargo_args.push("--manifest-path".into());
-            cargo_args.push(path.into());
-        }
-        if self.workspace {
-            cargo_args.push("--workspace".into())
-        }
-
-        cargo_args.extend(self.package.iter().map(|pkg| format!("-p={pkg}").into()));
-        cargo_args
-    }
-}
-
-/// Leave it for Cargo to validate these for now.
-impl ValidateArgs for CargoArgs {
-    fn validate(&self) -> Result<(), Error> {
-        Ok(())
     }
 }
